requirements/base.txt
django-environ==0.4.5
psycopg2==2.8.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
